Test FunctionK in scala 3
